backend/requirements.txt
flask
tensorflow
opencv-python
pillow
numpy
pymysql